cpDep.sh: Greatly enhanced the quality of the function safeCopyFile.
Added important comments to tell how the different sections of that function
works. Also made it correctly handle the difference in time creation of
the destination compared to the origin. If there is a difference, we
copy the new version to the destination. There was a big bug that
completely rendered key conditionnals inoperative. This commit fixes all
that.