newJail: Added the new internal function cmkdir
This function behaves in a similar manner as 'mkdir -p' but
with a small distinction. The application mkdir when asked to change the
permission of the inherent directories actually does so _only_ on the
last child of the entry. We actually want to apply the permissions to
all the directories created and this is what cmkdir does.