newJail: Added a new user custom script.
The user custom script called rootCustomConfig.sh is where the user
needs to put their custom changes/configuration to the jail.
startRoot.sh will no longer need to be customized like before. This
makes it easier to update the script without overwriting user configs.

Support was also added to run the jail from any path, previously it was
a requirement to be inside the jail's directory to start it correctly.