newJail: Rewrote the way the bridge is handled.
The bridge used to be implemented on the base namespace and masqueraded
to the net. This was removed and now it's the external interface that
is being masqueraded to the net. New options were added like the
possibility to join a bridge and to create one at the same time.
Creating a bridge is totally different now, it is strictly created
inside the jail's namespace and doesn't do anything as is. It's to
permit external jails to attach to it only. But it does not relay to
the internet anymore.