newJail: Removed the requirement for root.
We don't need to create any device nodes when using busybox so we remove
their creation. Also moved ownership changes (which requires root) to
when the jail is started (which _absolutely_ requires root as it is
chrooted). Without these, everything done by newJail.sh is now
unprivileged and it thus no longer requires root access.