newJail.sh: startRoot - Rewrote the function runChroot.
This function now supports the '-root' flag to run the chroot with root
access. It also better supports customized commands. The idea is to
make it as customizable as possible so it can be used when more complex
commands are to be used. This is managed by better handling the flags
thus function gets (it makes sure they maintain their coherence better).

Also paving the way to future User namespace support.

Also removed the function runShell completely as it is no longer
necessary thanks to the newest changes to runChroot.
The code of runShell was now moved to the shell command in
rootCustomConfig.sh.