newJail: startRoot - Removed the function startChroot.
Also made the function prepareChroot return a status rather than
exiting and made the calling functions exit on error.

This change is to make it clearer to the end user on how to use the
different parts of startRoot. And especially give them the tools to
make their own CLI commands.