newJail: Added support for additional entry shells.
Previously we could only have at most one active shell per jail. This
commit changes this and makes it possible to potentially support an
infinite amount of external shells per jail. The shell entry command
now starts a new shell in the jail if the jail is already running. It
will still start a new jail as previously if the jail is not running.

The new function findNS was created for looping and finding the child
namespace from which we can take the namespace pid. This pid is then
fed into the function nsenter to enable this endeavor. We also
create the file run/jail.pid with the main script's pid which is the
basis from which additional entry shells can be implemented.