newJail.sh: StartRoot - stopChroot : Added better unmount support.
Now, it unmounts directories first when stopping. Previously it would
unmount last and if there was any error unmounting it would remove
means to figure if the jail was closed correctly. This way, when there
is an unmount error, you can do jt stop  and it will clean everything
up only if the unmounting were done correctly.