newJail.sh: Added support for starting a jail in daemon mode.
You have to start it with the super script : sudo jailtools daemon
doing this directly with startRoot.sh will amount in an unresponsive
hanged session which is hard to kill. There is an error message that
is shown to teach the user not to use this directly. To get rid of it,
just kill it with -9.