newJail.sh: startRoot - Fixed joining started jail.
Fixed a very important intermittent bug where sometimes the jail joined
would not be of the current namespace. It turns out that this was all
related to the pid we searched. If the pid we saved was part of an
existing longer pid, we risked the chance of it being chosen as the
correct pid. This fix makes sure this doesn't happen and only the pid in
question in being chosen.