Added support for adding devices on the fly.
This is a major change because the jail's /dev directory is now being
mounted as a tmpfs directory. This means that it's content is flushed
ever time the jail is stopped. This was added so jails can be easier to
backup. As before this step required root access. To add devices to the
jail, you now have to use the addDevices function inside the prepCustom
function. Documentation was added to that function for how to do this.