reverted last commit and make the backup happen no matter what.
The second patching strategy was actually a failure. Even though it
managed to actually patch and the output seemed right, it in fact wasn't
right at all. The new content of the script was not being added; only
the user's content was preserved... which is not at all what we want.
We want the new content to be the basis upon which the user content
is added and this is why the last patch was reverted. We prefer to let
the user correct the issues manually in that case.

Now, even if the patch fails, all the patching and all files are moved
to the backup location. We notify the user of the path of that backup so
they can attempt the upgrade manually if they want.