Changed the way startRoot works and other core methods.
Renamed the current startRoot.sh to jailLib.sh and made it into a
library. The new startRoot.sh contains only the entry point of the
process (which was in rootCustomConfig.sh). Now, rootCustomConfig.sh
has configuration variables to set the commands to run for each
jailtools command. The script jailUpgrade was enhanced greatly and
modified to support the new implementation.