Added bounding-set from setpriv and implemented its use.
This includes a patch which adds the feature to busybox and also the
setpriv command is set to be compiled in busybox.

Currently, all but the net_bind_service (open ports in the privileged <
1024 range) is allowed. All the other capabilities are dropped.

Eventually, we may want to allow the user to configure which
capabilities are to be available in the jail.