startRoot: Modified how arguments are handled by jailtools.
rootCustomConfig has the command sections in which it is possible to add
what commands are run when the daemon, start or shell commands are
called. It is quite annoying if say you have a command on the shell
command but can't just call a simple shell to test a jail. This patch
fixes this by giving the priority to the arguments passed to jt
<command> rather than the command in rootCustomConfig. So, if you had a
command set in rootCustomConfig and you wanted to get a shell instead
you would do : 'jt shell sh'   (it will supersede the command set in
rootCustomConfig and call a simple shell). Running 'jt shell' will still
run the shellCommand from rootCustomConfig as before.