jailLib: fixed userCreds variable.
This variable contains the user credentials of the jailLib.sh script so
even when the script is run as root the actual user that owns the script
will set to which user the jail will be started. The bug was that we
used $0 before which didn't quite work as it sometimes just gave a
generic 'sh'. So we now use the variable ownPath along with the
hardcoded name for the script to populate the variable.