Implemented rootDefaultConfig which holds the default configuration.
This file contains the default configuration and documentation for these
configs. It lifts the need to assign each of these configuration
variables from rootCustomConfig. As such, we can now leave just a
minimal undocumented subset of the configuration in rootCustomConfig.

At this point, the user will be able to just add these config variables
to rootCustomConfig and it will change the default value.

Removing a configuration variable will now just make it have the default
value.