Reworked the internet access method.
setNetAccess used to be used for privileged and unprivileged accesses
in different manners. This was removed and a new configuration called
disableUnprivilegedNetworkNamespace was added for enabling internet
access to an unprivileged jail.

Also, there were a lot of core changes to the networking flags. The
process was made more efficient and easier to maintain.