Fixed a bug where the config ipInt was not being set correctly.
the file rootDefaultConfig set the value of ipInt as a script which uses
extIp as the basis. The issue here is that this always based the value
on the default extIp and never in the one in rootCustomConfig so we
had the issue that extIp and ipInt were not in the same network (and it
is vital that they are).

This fixes the situation.