Added a new configuration toggle : realRootInJail
It permits to keep the real privileged user inside the jail (root). This has to
be used to host services like sandstorm which rely on the ability to
create further user namespaces. This is not possible otherwise as an
unprivileged user because these are blocked by the kernel. They may be
blocked for a privileged user too though, so this has to be tested.