utils: populateFile - fixed a major bug with the function.
This is due to the recent changes to place configuration variables
between double quotes - populateFile was expecting the special place
holders to be at the end of the line and this would make most of these
place holders not to be replaced by their respective values. This patch
fixes this.