Refactored part of the making process.
Musl and Busybox are now being built out of tree in a folder
corresponding to each inside the new 'build' directory. This has the
effect of no longer having busybox in the usual busybox/busybox folder.

To better support busybox path, a new generated script was created
called scripts/paths.sh which contains the busybox path. All the scripts
that use the 'bb' variable use it to define this variable with the
correct path. This had the unfortunate effect that old jails would no
longer work correctly before they are upgraded so a link was made to
busybox/busybox to permit old jails to upgrade to the new method.