Embedding all scripts into the 'jt' superscript.
This has the advantage of greatly simplifying the embedding into busybox
and it also makes for simpler code. The deployed 'jt' superscript is now
busybox itself and calling 'jt busybox <args>' is used to call the
internal busybox itself.

A lot of code was simplified because of this new deployment method.
Deploying 'jt' as a shell script is no longer supported.

Do note that we install as just 'jt' now, no longer as jailtools or
jtools like before.

There are a lot of other changes to properly convert existing code to
the new methodology.