jailLib: With this, sandstorm should work correctly.
It is necessary to use the realRootInJail and start the jail privileged
to make sandstorm work. Sandstorm is supposed to work without
privileged access but it relies on the user namespace which can't be
nested inside a chroot (which is what jails are at the base, a chroot).
There may be a way to start sandstorm without real root access in the
future.