jailLib: Now raising an error when we can't find the default route.
This is only valid on 2 conditions : If the setNetAccess is set to true
and if the netInterface is set to auto.

We'd expect a jail not to start if it needs internet access but it can't
get it. A started jail without internet access is not really useful
anyway. Hence this patch.