utils: jailStatus : Now also check the presence of jail.pid.
We used to only check the presence of ns.pid to figure if a jail was
running or not. This is not enough as ns.pid is just the inner jail
running. In the case of an error, it's usually the jail.pid (the pid of
the jailLib.sh process) that is missing. This is why we need to check
both.