startRoot: Added a better prepareChroot error handler.
Previously, we just would exit with the code 1. This is not really nice
because it could leave the inner process running (in ns.pid) and that
could lead to a zombie process. So when there is an error in the
prepareChroot function, we attempt to clean up the jail as best we can
before exiting with the code 1.