Simplified the install script and separated the embedding part.
Made the new script embedJT.sh which is the internal method to embed jt
into busybox before the compilation process. This all was in install.sh
which had a double function : this internal embedding in busybox and
permitting to install jt to some directory for the end user.

To make it easier to figure out, maintain and manage, these 2
functionalities were split. This drastically simplifies the install.sh
script and keeps the internal embedding stuff in embedJT.sh.