Added a better versioning scheme.
This versioning scheme uses git to get a short version of the current
commit hash and appends it to the current version. If the current
version is the release version, it does not append the commit hash to
the version.

This also automates the internal version assignment so it is no longer
needed to change the version in jailtools.template.sh.