getVersion: Fixed the script to better handle the inputs.
Thinking about it, we rely on busybox's awk for this but at this point
it is possible that the end user don't have it installed on their
system. So we have some kind of a chicken and egg problem that has to be
solved. We may have to check the availability of busybox as the normal
awk could lack features that we use. We could maybe use 'sub' rather
than 'gensub' in awk to ensure better backward compatibility too.