Fixed a major bug that happened if busybox wasn't already installed.
Please note that busybox is not supposed to be installed for 'jt' to
work correctly, hence why this is called a major bug.

The jailtools superscript 'jt' is embedded in busybox and it needs
access to a certain amount of commands to do it's bidding. Now, this
means that the jt executable is in fact busybox but with a 'jt' cloak.
It is however possible to call the inner busybox by using the 'exec -a'
argument (which is available in the busybox version of 'exec') to change
the argument 0 which is the basis upon which busybox is built upon to
provide links to the various applets it is built with. It turns out that
we had 2 major bugs in our current implementation.

The first being that even though busybox was the one calling the jailtools
superscript, the builtin commands (like 'exec') were called based on the shebang
parameter which was set to '#! /bin/sh'. This made the system use that
shell to provide the builtin commands rather than the inner busybox
(hence why this patch removed this line).

At that point, tests were attempted in a context where absolutely no
commands were available on the base system (we used a jail to provide
this of course ;) ) and a big amount of commands were simply not being
called through the inner busybox as they should have been.

The next few amount of patches will remedy this situation so as to
provide a fully self reliable solution as this is supposed to be.