embedJT: Added a function to check if an embedded file exists.
Also changed the runner and shower functions to make use of that new
function to properly state if it exists or not. This was required
because previously, any error in the script we called would end up being
flagged as a "file not found" error even though it clearly was not the
case.