jailLib: Separated busybox from internal to external.
when we use execNS (to run commands inside the jail's namespace), we
used to use the 'bb' variable which is the external busybox. Now, we use
another variable called 'nsBB'. In theory, with this version we could
still use the variable 'bb' and it would still work, but it definitely
would not work in the upcoming pivot_root branch, which is why the two
were separated.