jailtools: cd to the jail in question before doing commands.
This is for start/stop/shell and also daemon. This all stems around the
method we use to figure if the jail is running or not. We check the
jail.pid file with pwdx and compare that with the path of the jail we
know. Previously, this actually set the path where the command was
called which could be anywhere on the system, not helping us at all.
This makes sure that we 'cd' to the jail before calling the command,
which sets the pwd correctly in the pid of jail.pid.