scripts/jailtools: Fixed an issue when jt could not find itself in PATH
Actually, this is not totally the issue, the issue was especially
present when we did something like :

PATH= jt busybox jt

now, in that case, we would have absolutely no information whatsoever in
the jailtools superscript to bootstrap the necessary paths. What was
done, was to use the exported variable JT_CALLER from the first 'jt'
instance so further 'jt' instances could know the context.

Of course, as is, this is silly, but if we use 'jt' inside the code of
jailtools (like what we're doing in the status command), this is in
effect what we are doing and this is what caused bugs.

This commit fixes that.