tests/daemon: Fixed a bug that made the tests fail.
This test was the timeout of 5 seconds when calling a new jail daemon.
This timeout is no longer necessary as jailtools now includes it's own
timeout when calling a daemon and gives out an error when it's not
started correctly.

This patch is also rewrites the whole daemon tests to use the new technique of
testing.