tests/netAccessSU: Test network namespace on shell reentry.
This is a bug in the current version where first creating a daemon and
then reenter it using the shell command would actually give the host's
network access rather than the jail's network namespace. This means if
the jail was supposed to not have internet access, it would actually get
it with the command shell. Do note that commands started through the
'daemon' command are not faulty, they respect the expected network
namespace.