config: Implemented support for config variable expansion.
This is to support having shell script code into configuration values.

Among others, this is to add support for the bridgeName, vethExt and
vethInt which take the jailName and caps it's length to specific size.
This actually makes it possible to get the resulting value of the
configuration rather than the shell script code, as it was before.