jailtools: Fixed a reliance on the path /proc/$$/exe.
We had an issue with starting a jail, in the initalization phase, the
path /proc/$$/exe was not correct so it was not reliable. This patch
changed the method and uses the command 'realpath' ínstead.

Also added a check to make sure we don't have any further issues in the
future.

This was a very very hard to debug and long standing bug in the jail
starting process and this effectively fixes the issue.