utils: detectJail - Added 2 new jail detection methods.
First is simply to check the path of the jail.pid process with pwdx but
it's not enough by itself. Next is the ultimate one which is to check
the process' mountinfo and check for one which has a '/root /' entry
path. We then use that entry with the detectJail function to see if it
is a valid jail or not. If it is, we are 100% that the jail is actually
running, even as root and especially with pivot_root
(which is trickier to figure out).