Added the new directMounts configuration to handle singleMounts.
This is the last piece that required prepCustom to stay in the prior to
pivot_root. With this out of the way, prepCustom was moved to the end of
prepareChroot (like it used to be before pivot_root era). Also added
support to having comments and empty lines removed from the list. (that
part will be used for other multiline configuration too, soon).

This all was actually to fix the firewall issues. Firewall rules have to
happen *after* the network devices are initialized and the "prior
pivot_root" part was way before that. That made any firewall rules added
to prepCustom fail automatically.

Eventually, I think it would be best to also have a special
configuration for firewall rules as well, so stay tuned for that.