startRoot: Removed stopOnError because that would not have the effect wanted.
The problem with it is that it would trigger a stop of a jail in the
situation where we attempt to start an already started jail. In that
case, we expect the second start attempt to fail and leave the jail
running. But due to stopOnError, it would actually stop the running
jail. So baring a better way to stopOnError, it was removed in this
commit.