tests/firewall: Converted the code to use jt --run jt_firewall directly.
Previously, we were making a front script to do the same thing in
test.sh (firewallFront.sh). But with the new changes with IS_RUNNING and
the firewall script, we can use jt --run jt_firewall now instead and
this is what this patch converted in all the firewall sub tests.