utils: Implemented 3 new functions to get the UID, GID and the combination of a user.
The user from which the credentials are extracted is the one that owns the file jailLib.sh in the jail root.
Maybe it would be better to use rootCustomConfig.sh at some point, if we
decide to forgoe with jailLib.sh in jails, but for now it's there and we
can use it for this purpose.