jailLib: Fixed a bug on setNetAccess=true privileged.
The firewall rules would not be properly cleaned up after setup because
the setting would be done incorrectly. The snat command normally gets
the ipInt value as a global variable but since we removed it, it would
no longer be able to do it's job. This patch explicitely sets the ipInt
and ipIntBitmask values to the snat command to fix the issue. We will
probably remove support for global ipInt and ipIntBitmask from the snat
command in the firewall module in the future as well.