firewall: Fixed a fatal issue with the internal firewall implementation.
It would no longer work because of changes with the variable nsBB and
execNS which previously would still run on the host system. But in the
newer version, execNS exclusively works inside the jail. To fix this
issue, we now use only the network namespace to call the binary iptables
from the host system. There was only marginal testing done with this
implementation so it may not work in all situations.