Completely changed how the firewall is being set in rootCustomConfig.
Now, there is a new value called firewallRules where the firewall rules
are to be set rather than in the special function prepCustom. Actually
that function (prepCustom) and stopCustom were removed completely.

At this point, it was no longer necessary to keep the firewallInternal
and External because we now access the firewall function directly
through the special specific syntax in firewallRules (which ends up
making it simpler for the end user). Also, in jailLib, the snat
(firewall rule) call was changed to reflect that change.