jailLib and utils: Implement the function getUtime which can be used for benchmark
This patch introduces that function for benchmarking the inner core initialization.
The benchmarking timing is available in run/innerCoreLog

It will be interesting to benchmark the other parts of the jail
initialization/starting process and eventually maybe use these as part
of the tests.

Maybe even have a special benchmark make rule which could be used to
assess the performances of new features.