utils: Added prepareScriptInFifo.
This function is a new core function which is used to access either
a direct file or an embedded file. With data passed into the fifo,
it is possible to run, import and do anything we could otherwise do
with a normal file.

It is also possible to run utils directly with IS_RUNNING=1 and
call the function directly that way. This way it is no longer necessary
to eval anything even to first load utils.sh.
You can run utils.sh prepareScriptInFifo to load utils.sh into a
fifo, load it and then have access to all the functions in utils.sh.