This is version 0.3.0.
Mainly, this version is the "removal" of the 2 main scripts in jails : startRoot.sh and jailLib.sh.
Those are used from inside the 'jt' superscript.
It still support a file version of the scripts, in which case 'jt' will detect and use them rather than the embedded ones.
It will however be the responsibility of the user to upgrade these files at that point.
We may make a mention when one of those differ from the current version to let the user know that they could upgrade them manually.

In order to do this, we create a fifo file in run which we call run/instrFile. This fifo
is piped into with the script (either startRoot.sh or jailLib.sh) and we can then run it with 'sh'.